Move sqlite_settings to sqlite.settings
