+++ qcodes/__init__.py
@@ -79,7 +79,7 @@ from qcodes.dataset.measurements import Measurement
 from qcodes.dataset.data_set import new_data_set, load_by_counter, load_by_id
 from qcodes.dataset.experiment_container import new_experiment, load_experiment, load_experiment_by_name, \
     load_last_experiment, experiments, load_or_create_experiment
-from qcodes.dataset.sqlite_settings import SQLiteSettings
+from qcodes.dataset.sqlite.settings import SQLiteSettings
 from qcodes.dataset.param_spec import ParamSpec
 from qcodes.dataset.database import initialise_database, \
     initialise_or_create_database_at
